refactor: calcNumbersWithSelectedOperation function in Operation Component
